# src/IO.Swagger/wwwroot/swagger-original.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Manipulação de livros e suas dependências ",
    "version" : "v1",
    "title" : "Livraria Virtual"
  },
  "host" : "virtserver.swaggerhub.com",
  "basePath" : "/IKokiri/Livrariav/v1",
  "schemes" : [ "https" ],
  "paths" : {
...
    "/generos" : {
      "get" : {
        "tags" : [ "genero" ],
        "summary" : "Busca generos do armazenamento",
        "operationId" : "getGeneros",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "success",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Genero"
              }
            }
          },
          "400" : {
            "description" : "não encontrado"
          }
        }
      },
      "post" : {
        "tags" : [ "genero" ],
        "summary" : "Adiciona um novo genero",
        "operationId" : "addGenero",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "dados do genero que será adicionado",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Genero"
          }
        } ],
        "responses" : {
          "405" : {
            "description" : "Dados inválidos"
          }
        }
      }
    },
    "/generos/{id}" : {
      "get" : {
        "tags" : [ "genero" ],
        "summary" : "Busca genero por id",
        "description" : "Retorna um unico genero",
        "operationId" : "getGeneroById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "identificador do genero de retorno",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "seccess",
            "schema" : {
              "$ref" : "#/definitions/Genero"
            }
          },
          "400" : {
            "description" : "ID inválido"
          },
          "404" : {
            "description" : "Genero não encontrado"
          }
        }
      },
      "put" : {
        "tags" : [ "genero" ],
        "summary" : "Atualização de um genero",
        "description" : "Efetua a alteração de um ou varios dados de um genero",
        "operationId" : "updateGenero",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "identificador do genero que será alterado",
          "required" : true,
          "type" : "integer"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "dados do genero que serão alterados",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Genero"
          }
        } ],
        "responses" : {
          "400" : {
            "description" : "Genero inválido"
          },
          "404" : {
            "description" : "Genero não encontrado"
          }
        }
      },
      "delete" : {
        "tags" : [ "genero" ],
        "summary" : "Remoção de um genero",
        "description" : "Remove um genero",
        "operationId" : "deleteGenero",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "identificador do genero que será removido",
          "required" : true,
          "type" : "integer"
        } ],
        "responses" : {
          "400" : {
            "description" : "Genero inválido"
          },
          "404" : {
            "description" : "Genero não encontrado"
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "Genero" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer"
        },
        "genero" : {
          "type" : "string"
        }
      },
      "xml" : {
        "name" : "Genero"
      },
      "example" : {
        "genero" : "genero",
        "id" : 1
      }
    },
...
  }
}